requirements.txt
# Core RAG Pipeline Dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.5
google-generativeai==0.3.2

# Vector Database
chromadb==0.4.15
faiss-cpu==1.7.4

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# Web Framework & API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
tiktoken==0.5.2
sentence-transformers==5.1.0

# Performance & Caching
redis==5.0.1
aioredis==2.0.1

# Monitoring & Observability
prometheus-client==0.19.0
structlog==23.2.0

# Phase 4: Advanced Features & Optimization
psutil==5.9.6
cachetools==5.3.2
joblib==1.3.2
asyncio-throttle==1.0.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Production & Deployment
gunicorn==21.2.0
supervisor==4.2.5
